seperated notification query function from main display component
