+++ components/Newsfeed/Newsfeed.tsx
@@ -1,30 +1,17 @@
-import { collection, query, where, getDocs } from "firebase/firestore"
 import ErrorPage from "next/error"
 import { useEffect, useState } from "react"
 import { useMediaQuery } from "usehooks-ts"
 import { useAuth } from "../auth"
 import { Col, Row, Spinner } from "../bootstrap"
 import { usePublicProfile } from "../db"
-import { firestore } from "../firebase"
 import {
   Header,
   HeaderTitle,
   StyledContainer
 } from "./StyledNewsfeedComponents"
 import { AlertCard } from "components/AlertCard/AlertCard"
-
-type NotificationProps = {
-  bodyText: string
-  court: string
-  header: string
-  id: string
-  subheader: string
-  timestamp: string
-  topicName: string
-  type: string
-}
-
-type Notifications = NotificationProps[]
+import { NotificationProps, Notifications } from "./NotificationProps"
+import notificationQuery from "./NotificationQuery"
 
 export default function Newsfeed() {
   const isMobile = useMediaQuery("(max-width: 768px)")
@@ -32,75 +19,39 @@ export default function Newsfeed() {
   const { user } = useAuth()
   const uid = user?.uid
   const { result: profile, loading } = usePublicProfile(uid)
-  const subscriptionRef = collection(
-    firestore,
-    `/users/${uid}/userNotificationFeed/`
-  )
 
   const [isShowingOrgs, setIsShowingOrgs] = useState<boolean>(true)
   const [isShowingBills, setIsShowingBills] = useState<boolean>(true)
 
   const [allResults, setAllResults] = useState<Notifications>([])
-  const [orgResults, setOrgResults] = useState<Notifications>([])
-  const [billResults, setBillResults] = useState<Notifications>([])
-
-  const [notificationsDisplay, setNotificationsDisplay] =
-    useState<Notifications>([])
-  const [shouldDisplayResults, setShouldDisplayResults] =
-    useState<boolean>(true)
+  const filteredResults = allResults.filter(result => {
+    if (isShowingOrgs && isShowingBills) {
+      return true
+    }
+    if (isShowingOrgs) {
+      return result.type === "org"
+    }
+    if (isShowingBills) {
+      return result.type === "bill"
+    }
+    return false
+  })
 
   const onOrgFilterChange = (isShowing: boolean) => {
     setIsShowingOrgs(isShowing)
-
-    setShouldDisplayResults(true)
-    if (isShowing && isShowingBills) {
-      return setNotificationsDisplay(allResults)
-    } else if (isShowing) {
-      return setNotificationsDisplay(orgResults)
-    } else if (isShowingBills) {
-      return setNotificationsDisplay(billResults)
-    }
-    return setShouldDisplayResults(false)
   }
 
   const onBillFilterChange = (isShowing: boolean) => {
     setIsShowingBills(isShowing)
-
-    setShouldDisplayResults(true)
-    if (isShowingOrgs && isShowing) {
-      return setNotificationsDisplay(allResults)
-    } else if (isShowingOrgs) {
-      return setNotificationsDisplay(orgResults)
-    } else if (isShowing) {
-      return setNotificationsDisplay(billResults)
-    }
-    return setShouldDisplayResults(false)
-  }
-
-  const notificationQuery = async () => {
-    let notificationList: Notifications = []
-    const q = query(subscriptionRef, where("uid", "==", `${uid}`))
-    const querySnapshot = await getDocs(q)
-    querySnapshot.forEach(doc => {
-      // doc.data() is never undefined for query doc snapshots
-      notificationList.push(doc.data().notification)
-    })
-
-    if (notificationsDisplay.length === 0 && notificationList.length != 0) {
-      setAllResults(notificationList)
-      setOrgResults(
-        notificationList.filter(notification => notification.type === "org")
-      )
-      setBillResults(
-        notificationList.filter(notification => notification.type === "bill")
-      )
-      setNotificationsDisplay(allResults)
-    }
   }
 
   useEffect(() => {
-    uid ? notificationQuery() : null
-  })
+    uid
+      ? notificationQuery(uid).then(notifications =>
+          setAllResults(notifications)
+        )
+      : null
+  }, [uid, setAllResults])
 
   function Filters() {
     return (
@@ -139,9 +90,9 @@ export default function Newsfeed() {
                     </Col>
                   )}
                 </Header>
-                {shouldDisplayResults ? (
+                {filteredResults.length > 0 ? (
                   <>
-                    {notificationsDisplay.map((element: NotificationProps) => (
+                    {filteredResults.map((element: NotificationProps) => (
                       <div className="pb-4" key={element.id}>
                         <AlertCard
                           header={element.header}
