+++ components/Newsfeed/NotificationProps.tsx
@@ -1,0 +1,12 @@
+export type NotificationProps = {
+  bodyText: string
+  court: string
+  header: string
+  id: string
+  subheader: string
+  timestamp: string
+  topicName: string
+  type: string
+}
+
+export type Notifications = NotificationProps[]
